requirements.txt
vllm[audio]>=0.4.0
azure-servicebus>=7.11.4